requirements.txt
flask
twilio
google-generativeai
groq
python-dotenv
firebase-admin
gunicorn